Figure update (#28)
* Add files via upload

* Added image of graph.

Added the image of the graph found in the /assets/ folder
(Line is:
![figure1]({{ site.baseurl }}/assets/2021-01-29-Exploring how to reduce memory consumption in metagenomics with K-mer graphs_files/2021-01-29-Exploring how to reduce memory consumption in metagenomics with K-mer graphs_1_1.png) )

* Added image directory.

This commit contains the image directory with a png file of the notebook's output from 2021-01-29-Exploring how to reduce memory consumption in metagenomics with K-mer graphs.

* Fixing LaTex rendering.

Using https://www.codecogs.com/latex/eqneditor.php to render LaTex equations.

* Line 39 formatted incorrectly.

* Again fixing line 39 formatting.

* Lines 49 and 50 formatting.

* Line 50 formatting.

* Lines 34 and 35 formatting.

Changing the formatting of these lines so they render $k$ as LaTex.

* Lines 34, 48 and 50 formatting.

* Formatting for hyperlinks on lines 34, 48 and 50.

* Formatting horizontal scrolling and line 198.

Reformatted the link in line 198 and the code such that hopefully now there's no horizontal scrolling.

* Line 36 formatting.

Line 36's link/LaTeX was not rendered correctly..

* Typo in line 32.

"...corresponds a..." should be "...corresponds to a..." in line 32.

* Correction of probability formula

The formula for the probability two nodes are connected was incorrect. Instead of 4^{k}/2, I had written 4^{k/2}. This commit should fix that along with consequences of the error. (E.g. in the code's formulas as well as the body of text.)

* Formatting.

Changing formatting of 4^k/2 in the paragraph starting one line 37 as well as final 2^k updated to 4^k.

* Line 44 error

Line 44 changed from 4^k to 4^k/2.

* Graph no longer correct.

This graph is no longer correct because of the update of the formula. The x-axis isn't the right scale.

* Updated graph.

This is a picture of the updated graph's output.

* Figure folder name adjustment.

The figure's tag did not link to the right directory.